refactor(plugins): extract computeRateLimitWarning to shared _rate-limit.ts
Same math will apply to Linear's rate-limit predictor as to GitHub's. Pure
functional util (remaining/limit/resetAt → warning string) — single impl
beats canonical-copy when both call sites need the same semantics.

`computeRateLimitWarning` grows a `tag` parameter so multi-API operators can
tell which budget is running out ("GH" today, "Linear" next slice). Github
call site in `GhPluginBase.observeRateLimit` passes "GH"; tests assert the
tagged shape.

## Human-Claude Interaction Log

### Human prompts (VERBATIM):
1. "/worker roost 492 AvesAlight/roost feat/linear-api-492 alex roost-worker-492 #roost-issue-492"
   → Worker context: implement linear-api slice 1.

(All other direction in this commit came via IRC channel #roost-issue-492 from
roost-lead-pm — the "extract to plugins/_shared/rate-limit.ts and import from
both" guidance is from that thread. Lead approved the flat `_rate-limit.ts`
sibling shape over a directory split that would have collided with the
existing `_shared.ts`.)
