+++ src/orchestrator/plugins/github/__tests__/plugin.test.ts
@@ -4,7 +4,7 @@ import { GitHubIssuesPlugin } from '../issues-plugin.js'
 import { GitHubNewIssuesPlugin } from '../new-issues-plugin.js'
 import { GitHubCommitsPlugin } from '../commits-plugin.js'
 import { GhPluginBase } from '../base.js'
-import { RATE_LIMIT_WINDOW_MS } from '../github-api.js'
+import { RATE_LIMIT_WINDOW_MS } from '../../_rate-limit.js'
 import type { OrchestratorConfig } from '../../../config.js'
 import type { Command } from '../../../dispatcher-dm-handler.js'
 import type { PrSnap, IssueSnap } from '../types.js'
